styles.ts added
